bmad-dash/requirements.txt
rich>=13.0.0
textual>=0.47.0
GitPython>=3.1.0
PyYAML>=6.0
pytest>=7.0.0
